\name{DBH}
\docType{data}
\alias{DBH}
\title{Trees height and diameter at breast height}
\description{
The DBH data contains the diameter at breast height (dbh), height and condition data for all trees centered in 108 plots of size 0.2 hectare immediately following a single prescribed burn and also following three 5-yr (year) interval reburns (four burns total) and a single 15-yr interval reburn (two burns total) and associated treatment information. The trees information were established from mixed ponderosa pine (Pinus ponderosa Dougl. ex Laws.) that contained scattered western junipers (Juniperus occidentalis Hook.). The
plots were located in the Malheur National Forest on the southern end of the Blue Mountains near Burns, Oregon, USA.
}
\usage{data(DBH)}
\format{A text file with 5732 observations from 17 variables related of trees characteristics such as dbh and height.}
\encoding{latin1}

\references{
B. K., Kerns, D. J., Westlind, and M. A. Day. 2017. Season and interval of burning and cattle exclusion in the southern blue mountains, oregon: Overstory tree height, diameter and growth. Forest Service Research Data Archive, <doi:10.2737/RDS-2017-0041> .
}
\keyword{datasets}
