\name{HW}
\docType{data}
\alias{HW}
\title{Mixed norther hardwood}
\description{
Tree list from a U.S. Forest Service Forest Inventory and Analysis (FIA) plot PLT_CN 247006253010661 measured in 2012 and represents a typical northern hardwood uneven-age forest.
}
\usage{data(HW)}
\encoding{latin1}
\references{
\url{https://www.fia.fs.fed.us/library/database-documentation}
}
\format{
  A data frame containing 25 trees (rows) and two columns. Columns are the trees' scientific name and diameter at breast height in inches.
}
\keyword{datasets}
