\name{SW}
\docType{data}
\alias{SW}
\title{Southern loblolly pine plantation}
\description{
Tree list from a U.S. Forest Service Forest Inventory and Analysis (FIA)
plot PLT_CN 259082471010854 measured in 2011 and represents a typical
loblolly pine plantation.
}
\usage{data(SW)}
\encoding{latin1}
\references{
\url{https://www.fia.fs.fed.us/library/database-documentation}
}
\format{
  A data frame containing 18 trees (rows) and two columns. Columns are the trees' scientific name and diameter at breast height in inches.
}
\keyword{datasets}
